[SQUEAMISH-126] : passing to multiple light source

+++ Client/Views/GuiComponents/Simple/WorkBarComponent.cpp
@@ -69,6 +69,13 @@ namespace Views
 				SQ_CLIENT_ERROR("Class {} in function {} : Physics service is not referenced yet", __FILE__, __FUNCTION__);
 
 			}
+
+			m_shader_service = container->GetReference<Services::ShaderService>();
+			if (!m_shader_service)
+			{
+				SQ_CLIENT_ERROR("Class {} in function {} : Shader service is not referenced yet", __FILE__, __FUNCTION__);
+
+			}
 		}
 
 		tabs_size.push_back(ImVec2(0, 250));
@@ -167,16 +174,18 @@ namespace Views
 
 				ImGui::BulletText("Directional light");
 
-				if (m_state_service && m_state_service->GetScene() && m_physics_service)
+				if (m_state_service && m_state_service->GetScene() && m_physics_service && m_shader_service)
 				{
 					is_there_light_directional_source = m_state_service->GetScene()->GetIsThereDirectionLight();
 					if (!is_there_light_directional_source)
 					{
 						if (ImGui::Button("Add directional light source"))
 						{
-							m_physics_service->RemoveLightSources();
 							m_state_service->GetScene()->SetIsThereDirectionLight(true);
 							m_state_service->GetScene()->SetDirectionLight(glm::vec3(0.f, -1.f, 0.f));
+							m_physics_service->RemoveLightSources();
+							m_physics_service->SetLightSourcesGeneralParameters();
+							m_shader_service->PassLightsParametersToSSBO(m_physics_service->GetLigthSources());
 						}
 					}
 					else
@@ -435,13 +444,15 @@ namespace Views
 	{
 		if (selected_renderer)
 		{
-			if (!is_there_light_directional_source && m_physics_service)
+			if (!is_there_light_directional_source && m_physics_service && m_shader_service)
 			{
 				bool is_light_source = selected_renderer->GetIsALightSource();
 				if (ImGui::Checkbox("Is a light source", &is_light_source))
 				{
 					selected_renderer->SetIsALigthSource(is_light_source);
-					m_physics_service->SetLightSources();
+					m_physics_service->SetLightSourcesGeneralParameters();
+					m_physics_service->SetLightsAttenuationsParameters();
+					m_shader_service->PassLightsParametersToSSBO(m_physics_service->GetLigthSources());
 				}
 			}
